Refactor generation of region mappings for active cells
This commit introduces a new helper function, 'regions', that
encapsulates the generation of region mappings for active cells
(e.g., the equilibration and PVT regions).  This promotes overall
readability of the main function body.

As a further aid to readability, create a simple macro for
extracting the *active* regions (those regions that contain any
active cells) and use that when iterating the active equilibration
regions.