Guard against negative initial oil saturations
In a three-phase problem (O/W/G) for which the water transition zone
extends into the gas transition zone we often experience negative
oil saturations when we define So=1-(Sw+Sg).  The reason is that
both water and gas are mobile in the same cell while the other parts
of this code expecs that either water or gas is mobile.

This refinement recalculates the initial water and gas saturation
distribution in such cells from a derived gas-water capillary
pressure defined as

   Pcgw(Sw) = Pcog(Sg=1-Sw) + Pcow(Sw))

and sets the initial oil saturation to zero in these cells.  We
furthermore assume that GWC=GOC (i.e., that the gas-water contact
coincides with initial gas-oil contact) in those regions that are
influenced by the modified saturation definitions.

This refinement is inspired by the ECL initialisation algorithm.