WCONINJE: Guard against empty keyword data
In strange, but nevertheless possible, cases it may happen that the
keyword data is empty.  This is the case when, for instance, the
operator chooses to disable (comment) all keyword records while
still leaving the actual keyword itself still enabled.

Prevent input errors by only converting data to DOUBLE if there
actually is any data to convert.  This refinement enables inputting
the original 'BC0407.DATA' deck.