Implement '~' expansion for user's HOME directory
This follows the example of FOPEN which, on Unix platforms, supports
the common shell behaviour of interpreting '~/' as an alias of the
current user's home directory.  The implementation is simplistic and
assumes that the environment variable HOME contains the actual
pathname of said directory.

A secondary benefit of this change is that we avoid false negatives
in 'prune_modules' and guarantee that the PATH contains fully
expanded absolute pathnames only.