Restrict VCS special pathname matching
Version control software special directory names like '.git' or
'.svn' are special only if the string occurs immediately following a
FILESEP.  Users should be allowed to name their directories

    project.git

and 'mrstModule' should recognise that as a valid module name.  This
change restricts VCS directory name pruning to the case of *actual*
special cases.

This feature existed in the original edition of 'mrstModule', but
was lost during the transition to the 'mrstPath' facility.