Refactor generation of 'fluid.info' string
This commit refactors and generalises the generation of the
informational string 'fluid.info'.  Specifically, we introduce two
new helper functions 'densstring' and 'mcstring' that handle,
respectively, the density-dependent and miscible/compressible
aspects of the information strings.

Overall, this simplifies the logic of the 'summarise' helper and,
moreover, introduces support for multiple PVT regions.