Split source term calculation into contribution and assembly
This commit splits the generation of the final sparse transport
source-term vector into two stages, calculation of individual
contributions and final assembly into (number-of-cells)-by-1 vector.
This change enables solving single-phase transport problems for
which we only require total rates (and affected cells) in addtion to
two-phase transport problems that require "water" (phase 1) rates of
injected fluids.

Effectively, function 'computeTransportSourceTerms' is split in half
and now returns a structure rather than a (sparse) vector.  That
structure can be passed on to new function 'assembleTransportSource'
which will compute the original sparse vector based on that
information.

Crucially though, the caller may choose to ignore the compositions
of injected fluids (i.e., field '.compi' for wells and '.sat' for
'src' and boundary conditions) by setting new option 'use_compi' to
FALSE (default TRUE).  In that case 'computeTransportSourceTerms'
will not reference the saturation fields and the caller can solve
"transport" problems that don't include saturations.

The solution is still not general, however, because function
'assembleTransportSource' assumes that it is being used in the
context of a single- or two-phase transport solve only.  Moreover,
we may have to further split the calculations to explicitly separate
boundary contributions from wells and 'src' if this is to become
fully applicable to time-of-flight calculations.

Finally, update callers accordingly:
  - twophaseJacobian
  - explicitTransport -- for time step estimation purposes only