Allow getRegMap to take ADI or empty val input.
If an ADI object is passed to getRegMap, the size of the
object need to be handled separately.
If an empty val is given, it is assumed that the
size of the val matches each case, and the reginx is found.

Also, allowing cellInx to be used even though number of
regions is only one.