Correct misleading documentation on refinment criteria
The documentation stated that the refinement function would
apply to blocks that violate the criterion

    block flow >= average model flow

meaning blocks with too *little* flow would be refined.
This is not what happens, so state the correct criterion
for reader reference.

While here, also correct a singular/plural mismatch in
the description of the controlling parameter, NU.