Fix detection of absent phases
Commit 7457fd5 modified the low-level phase-property constructor to
always return a 1-by-3 cell array of entities.  As a side-effect of
that change, we also changed the semantics of the return value for
absent phases.  Previously, the corresponding entity would be an
empty array but the new value is a cell array *containing* an empty
array.  Consequently, the check based on

    cellfun(@isempty, pvtfun)

in function 'blackOilPVT' would produce a false positive and thus
conclude that absent phases were actually present in the scenario.

This change-set introduces a more elaborate and robust check that
only references the phases that are declared active in the scenario
and explicitly clears the entities that are not declared active.
This, in turn, makes it possible to continue using the 'isempty'
condition as an arbiter within function 'blackOilPVT'.