Extend object selection and appearance control
This commit enables selectively attaching fault faces to the main
'block' and its neighbours for additional visual effects.  Fault
faces are plotted in a gray colour (RGB = REPMAT(0.7, [1, 3])) with
red edges to make them stand out.  We furthermore enable independent
selection of object transparency for the 'block', its neighbours,
and the fault faces of these objects (if enabled).

While here, switch the colour selection map for neighbours to a
brightened version of the COLORCUBE.

Finally, extract the underlying fine-scale grid from the coarse
grid's '.parent' rather than requiring that the grid be passed as an
independent parameter.  Update all calls in clients accordingly.