Use more direct construction of openShutFlag
The construction

    { comp{:,6} }

is needlessly opaque (and slow) because we construct a CS-list only
to immediately construct a (row/horizontal) cell array from that.
It is a lot easier to use

    comp(:,6)

which constructs the requisite (column/vertical) cell array
directly.  The latter construction also removes the need for the
transposition of the RLDECODE result.