mrstPath: Support implicit 'search' verb
This commit extends the behaviour of function 'mrstPath' such that
we implicitly perform a 'search' if we're NOT called with a known
verb.  This allows using the function as

    f = fullfile(mrstPath('deckformat'), 'data', 'spe1.data')

which is a very common (and attractive) use case.

While here, clean up the documentation of the 'search' keyword and
document the special case of single input/single output 'search'
returning a string rather than a cellstring.

Suggested by: @kalie