mrstPath: Exclude certain directories during 'addroot'
This commit introduces an explicit list of directory names that are
not considered for automatic registration in the module catalog.

Immediate subdirectories named

    (o) data
    (o) deprecated
    (o) experimental

(subject to platform specific conventions on case sensitivity) will
*NOT* be automatically entered into the module mapping.  If the user
needs modules with such names, the registration must be performed
manually using the 'register' verb.

This is an escape clause that allows putting additional directories,
possibly containing unfinished/incomplete features, into the modules
or repositories to preserve history but not automatically activating
these features for unsuspecting users.

While here, finish the documentation of the 'addroot' verb.

Suggested by: @kalie