mrstModule add: Support per-module activation script
This commit implements support for per-module activation (on-load)
scripts.  If there is a file called

    modload.m

in the module's top-level 'private' directory, i.e., if the file

    fullfile(module_root, 'private', 'modload.m')

exists, then we 'run' it as ordinary MATLAB code.  If the script
fails (i.e., calls ERROR), we don't activate the module but emit all
messages from the script to help the user diagnose the activation
problem.

This is in preparation of reworking the support for third-party
modules (e.g., the MatlabBGL), but we foresee additional uses for
'modload'.  One possible use of the feature is to allow a module to
recursively and automatically load its prerequisite modules.