Rework support structure for third-party modules
This commit introduces a general location for storing and managing
third-party modules.  The tentative location is

    fullfile(ROOTDIR, 'utils', '3rdparty')

with subdirectories denoting the individual modules.  As an example
and demonstration of the mechanism, we support the 'MatlabBGL'
package in a module named 'matlab_bgl'.

The module's activation script (modload.m) initially fails, but
advises the user to run the prepackaged download script
(downloadMBGL.m) to prevent future activation failures.  The script
downloads the software into the 'matlab_bgl' directory, makes it
suiteable for use with recent editions of MATLAB (by removing the
"assert.m" file) and patches the user's 'startup_user.m' script to
re-register 'matlab_bgl' with correct location.