MatlabBGL: Prune 'modload' function upon download
The module activation of module 'matlab_bgl' is a function.  As such
it gets put into MATLAB's built-in function cache once loaded.  This
meshes poorly with installing a new version of the file at the end
of the download script.  While arguably a bug in MATLAB (M ought to
check modification times of function files), the work-around is
trivial (just "clear" the function), so that the activation
instructions as advertised by 'modload.m.orig' continue to work as
advertised.