A simple Gustafsson-like selector.
Since we do not have any real error estimates for black oil time stepping, we are more than satisfied with saying that our "tolerance" is:
desired iteration count / max iterations
Which, in fact, turns out to work quite well. I implemented the fallback strategy uncritically from the paper and the use of log there may not be the best idea.