impesTPFA: Defer trans. calculation to caller
This commit moves the responsibility for calculating connection
transmissibilities from within the solver and out to the caller.
This increases the solver flexibility and makes it easer to use
upscaled transmissibilities, as well as to include multipliers like
ECLIPSE's MULT* keywords.

The flexibility comes at a price for new users.  In the typical
case, the caller must insert a statement of the form

    trans = 1 ./ accumarray(G.cells.faces(:,1), 1 ./ T)

in which 'T' is the one-sided transmissibilities computed by the
MRST function 'computeTrans'.  The result 'trans' must then be
supplied to function 'impesTPFA'.  We should probably create a
simple utility that handles the harmonic averaging for the most
common cases.

Update callers accordingly.