Transmissibility upscaling: Move convex optimization out main routine
This commit moves the local transmissibility upscaling algorithm,
opt.opt_trans_alg='local', out to a new "private" routine,
'localLinOptTrans'.  This code depends on a convex optimization
package, CVX, that isn't always available.  It is better to have
that feature in a separate function file so that it doesn't confuse
the Code Analyzer (a.k.a., MLINT).