merge_options: Relax type checking for cell arrays
This commit introduces a particular feature in the processing of new
field values.  Previously, function 'merge_options' would fail
(i.e., call ERROR) if the new field values was of a class/type
different from that of the existing field value.  In this commit we
relax that requirement in the special case of the existing field
value being a CELL array.  We now allow ANY type of new value to
override CELL arrays, but always guarantee that the result is a CELL
array.

This, in turn, allows clients (i.e., other MRST functions) to accept
hetereogeneous collections of values in a single option.  As an
example, consider a hypothetical function 'getData' that retrieves
data sets from an external source (e.g., a URL):

    function data = getData(varargin)
       opt = struct('dataSet', { {} }, ...
                    'other'  , @options);
       opt = merge_options(opt, varargin{:});

       if ~ isempty(opt.dataSet),
          data = get_data_for(opt.dataSet{:});
       end
    end

The user of function 'getData' is now able to use either of the
syntaxes

    data = getData('dataSet',   'X'  )
    data = getData('dataSet', { 'X' })

to specify retrieval of data set 'X'.  Furthermore, we can use the
syntax

    data = getData('dataSet', { 'X', 'Y', 'Z' })

to specify retrieval of all data sets 'X', 'Y', and 'Z'.  Additional
examples are listed in the documentation.

Suggested by: [at] moynar