readFieldFmt: Implement special treatment of zero-element vectors
This commit moves the special-case handling of data vectors with
zero elements out from within the type-checking SWITCH and ahead of
that SWITCH altogeher.  While not a particularly common case, it
does happen occasionally, for instance in the specification

    'LGRNAMES'           0 'CHAR'

The previous code would fail in this case, specifically in the
CELLSTR call, because 'number' would be zero whence FSCANF would
read nothing from the 'fid' and CELLSTR would complain about its
input not being a string.

Initial bug report and example data kindly provided by

    Ludovic Ricard, CSIRO