initEclipseModel: Compute and return connection trans
This commit changes the interface of experimental function
'initEclipseModel'.  Previously, we computed the one-sided
transmissibilities and deferred calculating aggregate connection
values to the caller.  In this change we also compute these
aggregate values, but allow deferred calculation if the caller so
desires.  The connection transmissibilities incorporate fault
multiplier values if present.