simulateScheduleAD: Verify linear solver before assignment
This commit verifies that the linear solver (opt.LinearSolver) is an
instance of class LinearSolverAD prior to assigning the value to the
solver's equally named property.  This changes the overall semantics
of function 'simulateScheduleAD'.  Class 'NonLinearSolver' is a
handle so cannot be modified without changing the callers view of
the object.  The new semantics (don't assign invalid linear solvers)
is safer from the caller's point of view.