Use compatibility function for "UNIQUE(..., 'stable')"
This commit switches all in-module uses of

    [c, ia, ic] = unique(..., 'stable')

to use the compatibility wrapper 'uniqueStable' in its place.  There
is no change in functionality for users that already have a UNIQUE
that supports the 'stable' option (MATLAB 7.14/R2012a or later), but
this allows users of MATLABs prior to that to run solvers based on
the automatic differentiation module.