Don't generate unused object.
This commit switches all existing calls to function 'uniqueStable'
to ignore its second return value (IA) which was unused at all call
sites.  Moreover, we do this in a way that's compatible with MATLABs
prior to release 2009b (7.9) which don't feature the special '~'
language feature to ignore particular return values.