Make 'mrstModule list' reflect active path
This commit reverses the order of the elements in the 'MODLIST'
array returned by the 'add_modules' helper.  This means that the
return value reflects the actual module directory order in MATLAB's
search path.

Previously, the commands

 mrstModule add deckformat ad-core ad-blackoil ad-fi
 mrstModule add deckformat
 mrstModule list

would produce the output

 Currently active MRST modules
   * deckformat
   * ad-core
   * ad-blackoil
   * ad-fi

despite the order in the search path being

 deckformat, ad-fi, ad-blackoil, ad-core

from beginning to end.

With this change, the output of "mrstModule list" is

 Currently active MRST modules
   * deckformat
   * ad-fi
   * ad-blackoil
   * ad-core

whence the listed order from top to bottom reflects the actual
location in the search path.  This used to be true earlier too, but
was lost in a previous change.