Added support for calling the new operator constructor.
The new operator constructor setupOperatorsTPFA in mrst-autodiff intends
to fix an inconsistency in the sign of the discretized 'div' and 'grad'
operators.  'setupSimCompVE' can now call this new constructor instead
of the old 'setupSimComp' if an optional argument is passed.  To
encourage migration to the new standard, a warning is displayed whenever
the old standard is used.