Added small check to allow function to run on geometry-less top-surface grids
Slightly modified twophaseJacobianWithVE_s so that its getFluxCap
function works as intended on top surface grids even when no computed
geometry is present, as long as transmissibilities are provided.  This
fix is relevant when using this function with grids with periodic
boundary conditions, for which computeGeometry cannot be directly applied.

No functionality is changed, we just avoid computing the 'C' matrix in
cases where it is not needed.