merge_options: Generate useful diagnostics for "tricky" callers
This commit introduces a measure of robustness in the generation of
diagnostic messages concerning missing or unsupported options.  In
the case of object methods, the "caller" string includes the class
name as a prefix with a dot (period) to separate the class from the
method name.

As an example, the constructor of class TwoPhaseOilWaterModel is
translated to the string

    'TwoPhaseOilWaterModel.TwoPhaseOilWaterModel'

Function WARNING then fails to interpret the resulting string as a
message ID and proceeds to emit it as the complete diagnostic.  This
in turn omits the other, potentially much more useful, information
about *which* option was unsupported in a particular call.

This commit replaces non-word characters in the "caller" string with
underscores whence WARNING *can* interpret the first input as a
message ID and emit all information in the diagnostic.