readDefaultedRecord: Tighten definition of 'defaulted' item
This commit makes the definition of what consitutes a defaulted item
in a record more strict.  Previously a well template string like

    PROD0*

would be matched as a default/skip item which would produce results
that were difficult, at best, to analyse.

This commit switches the definition of skip fields to those that
consist exclusively of digits followed immediately by a single
asterisk (i.e., '*').