readWellKW: Remove support for WELLOPEN
This commit removes the support for keyword WELLOPEN.  We used to
treat WELLOPEN as an alias for WELOPEN, but the data formats of the
two keywords are completely different so treating the former as the
latter will only produce parsing errors.  Keyword WELLOPEN needs a
dedicated reading procedure which we will introduce on an as-needed
basis.