readWellKW: Add well template support to WELOPEN
This commit introduces a new helper function, matchWells, that takes
on the responsibility of matching well names and well templates to
existing records (typically in WELSPECS or WCON*).  In this process,
also reimplement the WELOPEN processor in terms of mathcWells to
allow well template matching in the records.

Together, these changes allow inputting UNISIM-I "_original" data
files from Texas A&M.