merge_options: Add support for returning unused options
This commit extends the protocol supported by function merge_options
to return unused 'key'/value pairs in an extra cell array if the
caller so requests.  If the caller requests extra output, then we
will not emit diagnostic output for unsupported/unused options.

The practical consequence of this change is that it becomes easier
to define options in standard way in one function and then to pass
other options on to lower-level implementation functions using code
of the form

    opt = struct('some', 'option')
    [opt, varargin] = merge_options(opt, varargin{:})

    % ...

    z = f(x, y, varargin{:})

in which case the call to function 'f' gets those options that were
not used/matched in the 'merge_options' call.  Essentially, this
means that a function can choose to support the union of the options
of its lower-level implementation functions although some care must
be taken if a function supports the same option as a callee.  This,
in turn, means that it for instance becomes easier to wrap a solver
like 'incompTPFA' into a higher-level time-step solver in the case
of a splitting method or, say, an MsFVM basis function solver.

Suggested by: [at] moyner