plotFaces: Reimplement and document 'outline' option
This commit reimplements the 'outline' option in terms of the
extended 'merge_options' protocol offered in commit 9919aa0.  In
particular, we dispense with the explicit indexing into varargin and
prefer to let function merge_options peel off the 'outline' option
and its argument before we pass the remaining 'key'/value parameters
on to function patch.

While here, also document the support for the 'outline' option.  We
have shipped it in MRST releases since 1.1 in April of 2009 so we
should acknowledge its existence.

No functional changes.