Ensure that only a single flag or a flag for all interfaces are valid input.
Previously, a single flag would be interpreted as valid for only the
first cell. A similar bug existed where any number of flags would be
accepted as long as they did not exceed the number of interfaces.

The additional assert and logic does not impact the runtime significantly.
A brief test on SPE9 (9000 cells) indicates that this assert takes about
0.5% of the time spent in upstream calls.