contourAtlas: Return contour handles only when requested
This change modifies the return parameter protocol from function
contourAtlas to return the contour handles created by CONTOUR3 only
if the caller requests that information.  This in turn eliminates
needless display of contour handle IDs in the command window unless
the caller explicitly suppresses the output (see the function
interactiveTrapping for an example of this).