Guarantee active 'co2lab' in module's examples
This commit guarantees that the co2lab module is always active when
we run an example in the module.  This allows running those examples
with a mostly clean environment (i.e., having only appropriate
mappings in mrstPath) through navigating to the example directory
and invoking the example script by name.

Note: Function 'moduleCheck' cannot be used for this purpose because
the function is not avilable until the co2lab module is active.