upscaling: Catch up to changes in core
This commit chases the renaming of function 'solveIncompFlow' to
'incompMimetic' and, in the process, ensures that the upscaling
routines can be run with only appropriate module mappings in
mrstPath.

Also, while here, switch to using an explicit Peaceman index inner
product specification.  Relying on the default is too susceptible to
problems in the face of changing discretisations.