plotFaces: Force Renderer=OpenGL when needed
This commit forces the figure renderer mode (figure property
'Renderer') to 'OpenGL' on Microsoft Windows and Mac OS X when the
caller supplies a non-scalar FaceAlpha property or when the
FaceAlpha data value is less than one (i.e., when the graphical
output is being generated with transparency effects).

This is a first attempt at working around the embarrassing and
frequent MATLAB crashes that happen on the aforementioned platforms
when transparency is used.