Adding a function to ReservoirModel which allows adjustment of sat. updates so that they do not violate the step.
This could be useful for some cases (for example avoiding the immobile
water saturation changing in a three phase problem). This issue was
suggested by hnil.

Note: No saturation updates use the function yet as I need to run more
tests before I will enable it in the main branch.