Exclude selected paths from "git archive"
This commit introduces a work-tree attribute list (see option
"--worktree-attributes" in git-archive(1)) that excludes the ignore
file '.gitignore', the attribute file itself and files whose paths
match '*experimental*'.

We also do not distribute the IMPES module (impes/).  It currently
does not work and restoring its utility is too much work.

This is in preparation of MRST 2014b.