msmfem examples: Draw partially transparent edges
This commit switches many of the examples in the msmfem module to
using partially transparent edges (EdgeAlpha = 0.125).  This is in
response to mrst-core's commit b5dedb6 which modified function
'plotCellData' to default to drawing edges where previously those
edges would be omitted.  With the subsequent change in mrst-core
commit 53fc668 to draw the edges in a grey colour ([0.4, 0.4, 0.4]),
using EdgeAlpha = 0.125 produces subdued edges which has a visually
appealing effect.