Add utility to check if two vectors represent same partitions
This commit introduces a new utility for checking if two particular
vectors actually represent the same partitioning of the cells in a
model.

Two vectors represent the same partition if either of the following
conditions hold

  - The vectors are exactly equal (according to ISEQUAL)
  - One vector is permutation/renumbering of the other--i.e., if
    each block of one partition has an exact equivalent in the
    other.

Under these criteria, the following snippet returns true:

    p = partitionCartGrid([9, 9, 9], [3, 3, 3]);
    i = randperm(max(p));
    q = i(p);
    isSamePartition(p, q)

Note: Function 'isSamePartition' is implemented in terms of sortrows
for all but the most trivial case (ISEQUAL(p1, p2)).  Consequently,
this should mostly be regarded as a develoment/debugging/diagnosis
tool rather than an intrinsic part of production code.