makeInternalBoundary: Refurbish documentation block
This commit reworks the documentation block to make it a bit more
complete and to properly distinguish the input grid from the result.
It is the user's prerogative to assign the result on top of the
input grid, but we should document the safe alternative by default.

While here, also rename the input face quantity from 'f' to 'faces'
throughout the documentation block to mirror the actual parameter
name of the function and correct a spelling error,

    interal -> inter*N*al

in the cross reference section to fix automatic hyperlinking in the
Command Window.