makeInternalBoundary: Duplicate nodes on copied faces
This commit extends the behaviour of face and node copying to also
introduce new nodes.  Specifically we introduce exact copies of the
affected nodes, including their coordinates from 'nodes.coords',
into the node list and assign the copies' node identities to the
duplicate faces.  This allows subsequent geometric grid splitting
(e.g., to represent a fracturing process) if the caller so desires.

We introduce a new helper function, copy_facenodes, to handle the
details of this copying and defer the direct index manipulations to
function 'blockDiagIndex' in the process.