readVector: Be more lenient in comment matching
This commit extends the comment marker matching to support arbitrary
amounts of whitespace characters prior to the comment marker itself.
That, in turn, allows putting a comment after the very last data
element of a vector that has a fixed number of elements (e.g., the
COORD keyword).

With this change we are able to input the COORD vector of the
UNISIM-I-H data set which ends in

    ... 3361.3901 -- 81, 58
    /

Without this change, function 'readVector' would produce a
diagnostic of the form

    Error using readVector (line 167)
    Unexpected input while reading keyword COORD.
    Found ' -- 81, 58' at position 393678.

with the leading space in the unexpected input pointing to the error
condition that is fixed in this commit.