Added a local, modified copy of stepOG from the old framework.
In order to make the code written in the old framework run after
variables have changed to water instead of oil, the step function
provided by the old framework is no longer compatible (stepOG updates
oil and gas, not water and gas).  Rather than modify the code in
mrst-autodiff, a local copy has been added, which updates water rather
than oil, and shadows the original one.  This is a temporary solution in
order to make the old code work at the moment (needed for comparison
with new code).   When everything has been ported, this can be removed.