convertWConInje: Support defaulted BHP upper limit
The unit converter for WCONINJ already supported an unspecified
(defaulted) BHP upper limit.  Install similar treatment in the unit
converter for keyword WCONINJE.  Note that

    readWellKW>readWConInje

uses 'NaN' in the input template so we need to use predicate ISNAN
rather "== -1" to identify defaulted BHP limits.

This is a robustness fix in preparation of MRST 2015a.