MatlabBGL: Chase new location
This commit fixes the automatic download script for 3rd-party module
'matlab_bgl' to reference the current location of the .ZIP file.  As
a fall-back we echo the FEX location in the command window in case
we fail to download the bits.

While here, also prune the original modload script.  This was
renamed into modload_fallback and not properly removed when the
mechanism was first designed.