Fix of shear thinning in sequential code.
Changes in polymer model to handle the shear thinning properly
in the sequential code. The correct thing to do is to store
the shear thinning multipliers in the pressure solver, and
then apply these multipliers in the transport solver.
The shear multiplier depends on both water flux and polymer
concentration, but we now consider them to be constant during the
transport step.

It was incorrect to fully remove the shear thinning from
the transport step as we tried before, because the water phase flux
was then weighted incorrectly, and thus the fractional flow was wrong.

Results from the SPE10 2D test case now show almost no difference
between the fully implicit solution and the sequential solution
(when using relatively small time steps at least).