Added mex version of interp1q.
This mex version is much faster than interp1q, which again
is much faster than interp1. The mex version interp1q_mex is
not used anywhere in the code as of now. It can be applied to
all fluid interpolations in the ADI code by making a hack in
mrst-core/utils/interpTable.m;

Replace the line
   yi = interp1(X, Y, xi, method, 'extrap');
by the line
   yi = interp1q_mex(X, Y, xi);

Then the mex version will be used instead. This shows a
significant speedup.