Moved 2 lines of code inside if statement to avoid error when code is run.
This is a work-around (hack) to deal with an empty "h.UserData" produced by the function "contour3()", which may be caused by the fact that MATLAB versions R2014 and onwards treat graphics handles as objects, not doubles. This issue may deserve more attention if h.UserData is required in other places.