Fix of relperm upscaling with anisotropic permeability.
There was an error in the relperm upscaling when the absolute
permeability was anisotropic. The upscaling of K*kr used
isotropic upscaling, and then the resulting K*kr / K become
rubbish. This is now fixed by using anisotropic upscaling
of K*kr.

Also, mimetic pressure solver is used as default pressure solver
in the absperm upscaling.