Add option to create base grids of Sleipner model
Base grids are created and written to .mat file only if option to do so
is specified and passed into function. These base grids refer to 1) the
caprock and bottom shale layers if available, and 2) the cut grid (sand
    layer only) before vertical-averaging or any refinement/coarsening.
The purpose of creating these base grids is to visualize the caprock and
shale layers.