ADI/sum: Always sum along columns (dimension=1)
This commit ensures that we're always summing the correct entries,
even in the unlikely event that the number of variables is one.
Previously, the statements

    [x, y] = initVariablesADI(ones([3, 1]), 2);
    sum(y)

would produce the result

    ADI with properties:
      val: 2
      jac: {[0]  [1]}

while the correct result, after this commit, is

    ADI with properties:
      val: 2
      jac: {[0 0 0]  [1]}

Note number of 'jac' elements.

While here, also apply the same correction to ADI/cumsum.

Signed-off-by: Bård Skaflestad <Bard.Skaflestad@sintef.no>