geom_2d*: Split face geometry out to separate helper
This commit introduces a new helper function, face_geom2d, that
handles task of computing the face geometry in two space dimensions.
This, in turn, reduces the lifetime of the 'edgeLength' intermediate
quantity to be just enough to compute the derived quantities,
thereby possibly conserving some memory in the overall process.