geom_2d*: Refactor sub-triangle area calculation
This commit introduces a new helper function, face_geom2d_subarea,
that handles task of computing the areas of the sub-triangles
made from the existing cell edges and connecting the edge nodes to
the hinge node.  This, in turn, reduces the lifetime of the
'cellEdges' intermediate quantity to be just enough to compute the
derived quantities, thereby possibly conserving some memory in the
overall process.