geom_2d*: Use single implementation in 2D and top-surface cases
This commit introduces a new helper function, geom_2d_impl, that
implements the 2D cell and face geometries.  This function is common
to both the original 2D grid and to the top-surface 2D grid with 3D
node coordinates.  The difference between the two is the definition
of the area of the quadrilateral that is spanned by the edge and
hinge node vectors and this implementation highlights this fact.