face_geom3d: Add support for computing on subset of faces
This commit introduces an API change in function face_geom3d.  We
now support computing areas, centroids, and normals on a subset of
the grid's faces--identified by a collection of face indices.  This,
in turn, is useful in the process of refactoring the cell loop into
a loop over blocks of cell indices.