geom_3d: Switch to on-demand calculation of face geometry
This commit moves the calculation of geometric primitives on faces
into the cell loop.  The main objective is to avoid creating large
intermediate arrays when computing the geometry of multiple cells at
a time.  As a particular benefit, we no longer need to create and
expose the 'localEdge2Face' accumulator.