geom_3d: move internals of cell loop to separate helper function
This commit introduces a new helper function, geom_3d_cell_block,
that contains the single-cell geomtry calculation of the

    for c = 1 : numC

loop in function geom_3d.  In the process, generalise the
implementation to apply to collection of cells rather than a single
cell only.  Also, introduce a new helper structure, 'geom', that
manages the intermediate and result data of geom_3d.  This helps
keep the parameter and return-value lists readable.

Isolated, this is a significant performance regression compared to
the original implementation of function 'computeGeometry', but it
lays the foundation for subsequent vectorisation.