geom_3d: Use verbose printing tailored to block processing
This commit updates the verbose-mode printing to reflect the block
nature of the processing.  In particular, we print the cell index
range (and total number of cells) currently being processed, as well
as well as an estimate on the processing rate (number of cells per
second).  Remove the face-section printing in the process because it
conflicts with the line oriented printing of cell blocks.