matrixBlocksFromSparse: Support non-square blocks
This commit adds support for extracting non-square blocks from the
main diagonal of a sparse matrix.  The existing syntax

    Sbd = matrixBlocksFromSparse(S, sz)

is supported for backwards compatibility and will behave according
to the original specification.  Extracting non-square blocks means
having to pass two vectors--'rsz' and 'csz'--of row and column sizes
respectively:

    Sbd = matrixBlocksFromSparse(S, rsz, csz)

As a convenience to the caller, we also support returning a vector
of start pointers that helps identify the actual numeric values of
an individual block.