Add utility for solving sequence of LLS problems
The new utility, leastsq_svd, solves linear least squares (LLS)
problems and its syntax and overall strategy is inspired by the
existing function 'invv' which explicitly inverts a sequence of
small square matrices.  We forward the actual numerical work to
LAPACK subroutine DGELSS and only take care of some book keeping in
the mexFunction() itself.

Syntax:

    x = leastsq_svd(A, b, rsz, csz)

with 'A' being the result of 'matrixBlocksFromSparse', 'b' the
concatenation of all LLS right-hand sides, 'rsz' being a vector of
row sizes and 'csz' being a vector of column sizes.

Note: This code is currently sequential only.