multiSymmEig: Enable concurrent solution of independent problems
This commit separates the process of solving a single eigenvalue
problem out to a helper function, solveEigenProblem().  This
function can the be scheduled as a separate OpenMP task.  We also
introduce two helper classes to identify the index ranges pertaining
to an individual eigenproblem.

Note that in order to avoid thread package incompatibilities, we
need to link to Intel's OpenMP Runtime Library which, on GLNXA64, is
available as

    fullfile(matlabroot, 'sys', 'os', 'glnxa64', 'libiomp5.so')

With this implementation I am able to solve 600,000 random symmetric
50-by-50 eigenproblems (eigenvalues and eigenvectors both) in
roughly 25 seconds using twelve threads.  The equivalent sequential
approach runs in roughly 280 seconds.