computeTrapVolume: correction in calculation
The trap height is the minimum of the cell height (Gt.cells.H) and
the difference between trap spill point elevation and top surface
elevation (i.e., trap_z - Gt.cells.z). This accounts for the possible
scenario where the formation's bottom in a trap is shallower than the
trap's spill-point depth.