Record Input: Support Quoted Substrings
This commit adds read support for records of the form

    1.0 1* 'H--' Z -1.234e5 '*PROD 1' 'MULTX-' /

which contain quoted substrings that would otherwise be interpreted
as multiple tokens or, in the case of 'H--', as introducing a
comment whereby the remainder of the record would be discarded.

This, in turn, is needed to support certain option settings in the
first record of the WELSEGS keyword.  This processing does make the
input reading more expensive.